[INTGW-951] Delete topic endpoint

+++ component/app-store-ui/src/app/forum/components/forum-main/forum-main.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from "@angular/core";
 import { AppState } from "../../../app.data.models";
 import { Store } from "@ngrx/store";
 import { Topic, GetTopicsParam, TopicResultPayload } from "../../forum.data.models";
-import { MatTableDataSource } from "@angular/material";
+import { MatTableDataSource, MatDialogRef, MatDialog } from "@angular/material";
 import * as forumActions from "../../forum.actions";
 import { Router } from "@angular/router";
 import * as globalActions from "../../../app.actions";
@@ -18,7 +18,10 @@ export class ForumMainComponent implements OnInit {
   public topics: TopicResultPayload;
   public searchQuery: string;
 
-  constructor(private store: Store<AppState>, private router: Router, private titleService: Title) {}
+  constructor(
+    private store: Store<AppState>, 
+    private titleService: Title,
+    public dialog: MatDialog ) {}
 
   ngOnInit() {
     this.store.select(s => s.forum.allTopics).subscribe(res => {
@@ -44,10 +47,35 @@ export class ForumMainComponent implements OnInit {
   }
 
   onTopicDelete(id) {
-    this.store.dispatch(forumActions.DeleteTopicAction({payload:id}));
+    const dialogRef = this.dialog.open(DeleteConfirmationDialog);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result=="delete") this.store.dispatch(forumActions.DeleteTopicAction({payload:id}));
+    });
   }
 
-  onTopicView(topic) {
-    this.router.navigate(["forum/view-topic"]);
+  openDialog(id): void {
+    
   }
+}
+
+@Component({
+  selector: 'delete-confirmation-dialog',
+  template: `
+<div mat-dialog-content>
+  <h2>Confirm Delete ?</h2>
+</div>
+<div mat-dialog-actions>
+  <button mat-button (click)="onNoClick()" cdkFocusInitial>Cancel</button>
+  <button mat-button mat-dialog-close="delete">Delete</button>
+</div>`,
+})
+export class DeleteConfirmationDialog {
+
+  constructor(public dialogRef: MatDialogRef<DeleteConfirmationDialog>) {}
+
+  onNoClick(): void {
+    this.dialogRef.close();
+  }
+
 }
